+++ Reach/Changes.md
@@ -1,5 +1,8 @@
 # Utopia Reach - List of Changes
 
+## v1.1.1 (in development)
+Supports **Flow v2.3**
+
 ## v1.1 - 27.09.2023
 This update continues the major refactoring process of v1.0 update.  
 Almost every component constructor is rewritten to support a new builder-style approach.