+++ Reflection/Changes.md
@@ -1,5 +1,8 @@
 # Utopia Reflection - List of Changes
 
+## v2.1.1 (in development)
+Supports **Flow v2.3**
+
 ## v2.1 - 27.09.2023
 Introduces a bugfix to **Stack** component management
 ### Breaking Changes